Move `utils.testing` to a directory
