+++ utils/testing/__init__.py
@@ -2,76 +2,16 @@ import ipaddress
 import json
 
 from django.contrib.auth.models import Permission, User
-from django.core.exceptions import FieldDoesNotExist, ObjectDoesNotExist
-from django.db.models import ManyToManyField
-from django.forms.models import model_to_dict as _model_to_dict
+from django.core.exceptions import ObjectDoesNotExist
 from django.test import Client
 from django.test import TestCase as _TestCase
-from django.urls import NoReverseMatch, reverse
+from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APIClient
-from taggit.managers import TaggableManager
 
 from users.models import Token
 
-
-def json_file_to_python_type(filename):
-    with open(filename, mode="r") as f:
-        return json.load(f)
-    return None
-
-
-def model_to_dict(instance, fields, api=False):
-    """
-    Returns a dictionary representation of an instance.
-    """
-    model_dict = _model_to_dict(instance, fields=fields)
-
-    # Map any additional (non-field) instance attributes that were specified
-    for attr in fields:
-        if hasattr(instance, attr) and attr not in model_dict:
-            model_dict[attr] = getattr(instance, attr)
-
-    for key, value in list(model_dict.items()):
-        try:
-            field = instance._meta.get_field(key)
-        except FieldDoesNotExist:
-            continue
-
-        if value and type(field) in (ManyToManyField, TaggableManager):
-            model_dict[key] = sorted([o.pk for o in value])
-
-        if api and type(value) in (
-            ipaddress.IPv4Address,
-            ipaddress.IPv6Address,
-            ipaddress.IPv4Interface,
-            ipaddress.IPv6Interface,
-        ):
-            model_dict[key] = str(value)
-
-    return model_dict
-
-
-def post_data(data):
-    """
-    Takes a dictionary of test data and returns a dict suitable for POSTing.
-    """
-    r = {}
-
-    for key, value in data.items():
-        if value is None:
-            r[key] = ""
-        elif type(value) in (list, tuple):
-            if value and hasattr(value[0], "pk"):
-                r[key] = [v.pk for v in value]
-            else:
-                r[key] = value
-        elif hasattr(value, "pk"):
-            r[key] = value.pk
-        else:
-            r[key] = str(value)
-
-    return r
+from .functions import json_file_to_python_type, model_to_dict, post_data
 
 
 class MockedResponse(object):
